fullstack-db-optimization-system-pro/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
asyncpg==0.29.0 # PostgreSQL async driver
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.4
fastapi-limiter==0.1.6
fastapi-cache2==0.2.1
python-dateutil==2.9.0.post0
gunicorn==22.0.0 # For production deployment
loguru==0.7.2 # Enhanced logging
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0 # For testing FastAPI apps
resend==0.7.0 # Optional: For sending emails (e.g., password reset, notifications)